chore: some refactors and fix attaching audio notes with only transcripts
s
